syncing first in block manager. seprating loops by aggreagator mode
