+++ block/produce.go
@@ -3,11 +3,9 @@ package block
 import (
 	"context"
 	"fmt"
-	"sync/atomic"
 	"time"
 
 	abciconv "github.com/dymensionxyz/dymint/conv/abci"
-	"github.com/dymensionxyz/dymint/settlement"
 	"github.com/dymensionxyz/dymint/types"
 	tmed25519 "github.com/tendermint/tendermint/crypto/ed25519"
 	cmtproto "github.com/tendermint/tendermint/proto/tendermint/types"
@@ -15,43 +13,9 @@ import (
 	tmtime "github.com/tendermint/tendermint/types/time"
 )
 
-// waitForSync enforces the aggregator to be synced before it can produce blocks.
-// It requires the retriveBlockLoop to be running.
-func (m *Manager) waitForSync(ctx context.Context) error {
-	resultRetrieveBatch, err := m.getLatestBatchFromSL(ctx)
-	// Set the syncTarget according to the result
-	if err == settlement.ErrBatchNotFound {
-		// Since we requested the latest batch and got batch not found it means
-		// the SL still hasn't got any batches for this chain.
-		m.logger.Info("No batches for chain found in SL. Start writing first batch")
-		atomic.StoreUint64(&m.syncTarget, uint64(m.genesis.InitialHeight-1))
-		return nil
-	} else if err != nil {
-		m.logger.Error("failed to retrieve batch from SL", "err", err)
-		return err
-	} else {
-		m.updateSyncParams(ctx, resultRetrieveBatch.EndHeight)
-	}
-
-	//FIXME: we can call to syncUntilTarget here instead of the loop
-
-	// Wait until isSynced is true and then call the PublishBlockLoop
-	m.isSyncedCond.L.Lock()
-	// Wait until we're synced and that we have got the latest batch (if we didn't, m.syncTarget == 0)
-	// before we start publishing blocks
-	for m.store.Height() < atomic.LoadUint64(&m.syncTarget) {
-		m.logger.Info("Waiting for sync", "current height", m.store.Height(), "syncTarget", atomic.LoadUint64(&m.syncTarget))
-		m.isSyncedCond.Wait()
-	}
-	m.isSyncedCond.L.Unlock()
-	m.logger.Info("Synced, Starting to produce", "current height", m.store.Height(), "syncTarget", atomic.LoadUint64(&m.syncTarget))
-	return nil
-}
-
 // ProduceBlockLoop is calling publishBlock in a loop as long as wer'e synced.
 func (m *Manager) ProduceBlockLoop(ctx context.Context) {
-
-	//PREVIOSOULY IT BEEN WAITFORSYNC
+	m.logger.Debug("Started produce loop")
 
 	ticker := time.NewTicker(m.conf.BlockTime)
 	defer ticker.Stop()
